requirements.txt
nose2
validators
tqdm
httplib2
ftfy
pandas
requests
pyyaml